remove achilles_cor and dependencies
related to issue #34
achilles_cor is used to generate the graph, as well as the graph_report. However, the information that achilles_cor provides for those two functions are are also included in the master_tables. Used make_top_table and *bottom_table to get gene lists, then format for graphs.