rename achilles to achilles_raw, and achilles_clean is now achilles
achilles is the raw dataset, but was called several times in app.R making a discordance between the raw data and my cleaned data. The raw data have a DOI (and therefore will always be available), and so I changed a few lines of code here to create the achilles_raw object, and then move to achilles for the remainder of the analyses. The saved file output is still achilles.RData, and was only moved to the end to make sure all of the data transformations were accomplished before saving the file.