Added validation steps to make more informative error msgs
This update addes validation steps to each of the shiny output objects. If not found, then will stop the operation and display the custom message. Also needed to change how the text was displayed on the plot page (beginning previous line 395) so that it would accept a validate() arguement. Fixes #41