lower memory request to 1G
Currently if we configure openshift to run two pods
for the website new deployments fail due to memory quotas.
This change will allow us to run two pods and gracefully
deploy new changes.

Based on openshift Memory metrics we are using ~ 1040 MiB.
This is just the requested amount and the app is allowed
to utilize more than this.