Adds deployment to redirect http://ddh.org -> https://www.ddh.org
Began with `oc new-app httpd --dry-run=True -o yaml > Redirect.yaml`

This provided an ImageStreamTag for the httpd image, pulled from the internal openshift docker registry, a DeploymentConfig to run a pod for it, and a service for incoming traffic.

I layered on a simple redirect config file to mount at /etc/httpd/conf.d/redirect.conf and added a route to accept traffic at http://datadrivenhypothesis.org

I started with nginx but wanted to keep it simple, and the provided nginx images are all s2i builds.