Add update_gene_summary.R
This commit adds a new script to run quarterly(?) which requires a gene_summary_raw.Rds file in the data directory. This is a copy of the current_gene_summmary.Rds file. Using this file, the code will add 4 new columns to this dataframe: pubmed_count, concept_count, and percent rank for each. These are static values, and the idea is to serve them either on the gene landing page, in the header of the similar/dissimilar gene dependency page, or both. These values will provide context for new data (columns) on these page which quantifies known relationships between genes via co-publication.