adds generate_pubmed_data.R
This script generates the pubmed data by downloading it from the pubtator ftp site, cleaning, and quantifying the co-occurance of two genes in a single pubmed document. This long file is then pivoted wider to fill empties with zeros, then pivoted back longer to complete an otherwise sparse matrix. This is computationally expensive (takes a while), but I think is a better option than handling and replacing missing values in the app.